libc/fildes: Implement seekoptab in struct regfile_tx
Move handling of seek ops from struct ofd_tx to struct regfile_tx.
